refactor(database): decompose the topology repository god-file by entity (WS-D)
Split internal/database/repository_topology.go (720 LOC) into role-based
files in the same package (repository_topology_*.go convention) — pure move:
node operations stay in repository_topology.go; link/target ops and
interface ops move to their own files.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
